+++ kernel/include/arch/x86/pic.h
@@ -6,15 +6,15 @@
 #define PIC2_COMMAND                0xA0
 #define PIC2_DATA                   0xA1
 
-size_t PIC_init();
-size_t PIC_setMask(
+uint32 PIC_init();
+uint32 PIC_setMask(
     uint8 a_irqLine
 );
 
-size_t PIC_clearMask(
+uint32 PIC_clearMask(
     uint8 a_irqLine
 );
 
-size_t PIC_maskUnusedIRQs();
+uint32 PIC_maskUnusedIRQs();
 
 #endif
