+++ util/kstdlib/src/kstdio.c
@@ -5,7 +5,9 @@
 
 #include <stdarg.h>
 
-size_t kprintf(const char *a_format, ...)
+uint32 kprintf(
+    const char *a_format, 
+    ...)
 {
     if (a_format == NULL)
     {
